Really ugly, but it handles the case last segment problem
If we can't read the next segment, we can't be sure what size we are
expecting to decompress. This creates a problem for how we handle
multiple compressed segments. The code in this commit works around that
as an edge case, but the implementation is really ugly. It is probably a
better idea to just check the value of the lseg_sz and react when it is
zero. For the case when there _is_ an next segment, we can use that
header value to double check that we decompressed everything in the
previous segment, but we probably shouldn't rely on it.

Related to this, we might be able to dump the 1MB malloc for cbuf and
dbuf and malloc/realloc based on the values we are finding in the
compressed extents. This should be easy for cbuf, but might be harder
for dbuf.

Signed-off-by: Ian Allison <iana@pims.math.ca>